完善测试代码

+++ Codes/HotelReserveSystem_Client/src/main/java/bl_Stub/hotelblservice_Stub/CommentOnHotelServiceImpl_Stub.java
@@ -57,7 +57,7 @@ public class CommentOnHotelServiceImpl_Stub implements CommentOnHotelService{
 	}
 
 	@Override
-	public boolean confirmComment(String username, int mark, String comment, String hotelAddress) {
+	public boolean confirmComment(String username, float mark, String comment, String hotelAddress) {
 		return true;
 	}
